requirements.txt
python-telegram-bot==22.6
websockets==16.0
aiohttp==3.13.3
